Removed 'Post integral' functionality from the package.
For a while I was working on code to make plots similar to those in
D. Post's 1995 papers, of the parallel heat flux that can be dissipated
by a SOL of a certain impurity content using radiation and conduction
from a lower temperature to some higher temperature.

See the paper 'Calculations of energy losses due to atomic processes in tokamaks with
applications to the International Thermonuclear Experimental Reactor'

Anyway I decided that this functionality, built on top of the fundamental
zero-D collisional radiative (with optional diffusion lifetime tau) calculations,
was not logically part of the package at the same level. Perhaps I will make another package
for those calculations that uses these as input though.